--- src/lib/index.js
@@ -1,5 +1,0 @@
-// aqui exportaras las funciones que necesites
-
-export const myFunction = () => {
-  // aqui tu codigo
-};
